Use NodeFactory to create and update nodes
Most ts.create* and ts.update* functions are deprecated as of TypeScript 4.0. The
current plan is to emit noiser and noisier warnings with each version
until the deprecated functions are removed in 4.3.

See this pull request for more details:
https://github.com/microsoft/TypeScript/pull/35282

Some method signatures and names have changed slightly.

The easy way to update is to use the global ts.factory object.
TypeScript transformations also have access to a NodeFactory in their
transformation contexts.

Note that we still use the deprecate getMutableClone function because
its replacement is not included in the TypeScript declarations. See
https://github.com/microsoft/TypeScript/issues/40507