Pass extra args from ts-migrate-full to subcommands, fix eslint error
There was a bug in which an eslint wasn't being seen. I think it's because looking for globs in an if statement behaves funny, or because that particular syntax isn't as portable as it could be. I used stat instead, which seems to work (my eslintrc.js no longer gets deleted, and if I delete it, an eslintrc is in fact added).

To make ts-migrate-full work with --sources, I just modified the script to pass any extra arguments along to the child commands it runs.